# gateway/cmd/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: GatewayOps API
  version: 1.0.0
  description: |
    GatewayOps is an enterprise MCP (Model Context Protocol) Gateway that provides
    secure, observable, and scalable access to MCP servers.

    ## Authentication

    All API requests require authentication using a Bearer token:

    ```
    Authorization: Bearer gwo_prd_...
    ```

    API keys can be created in the dashboard or via the API.

    ## Rate Limiting

    API requests are rate limited based on your API key configuration.
    Rate limit headers are included in all responses:

    - `X-RateLimit-Limit`: Maximum requests per minute
    - `X-RateLimit-Remaining`: Remaining requests in the current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

  contact:
    name: GatewayOps Support
    email: support@gatewayops.com
    url: https://gatewayops.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gatewayops-api.fly.dev
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /v1/mcp/{server}/tools/list:
    post:
      tags: [MCP]
      summary: List tools
      description: List all tools available on the specified MCP server.
      operationId: listTools
      parameters:
        - $ref: '#/components/parameters/ServerPath'
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      $ref: '#/components/schemas/ToolDefinition'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/mcp/{server}/tools/call:
    post:
      tags: [MCP]
      summary: Call a tool
      description: Execute a tool on the specified MCP server.
      operationId: callTool
      parameters:
        - $ref: '#/components/parameters/ServerPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool]
              properties:
                tool:
                  type: string
                  description: Name of the tool to call
                arguments:
                  type: object
                  description: Tool arguments
      responses:
        '200':
          description: Tool call result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCallResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/mcp/{server}/resources/list:
    post:
      tags: [MCP]
      summary: List resources
      description: List all resources available on the specified MCP server.
      operationId: listResources
      parameters:
        - $ref: '#/components/parameters/ServerPath'
      responses:
        '200':
          description: List of resources
          content:
            application/json:
              schema:
                type: object
                properties:
                  resources:
                    type: array
                    items:
                      $ref: '#/components/schemas/Resource'
  /v1/mcp/{server}/resources/read:
    post:
      tags: [MCP]
      summary: Read a resource
      description: Read a resource from the specified MCP server.
      operationId: readResource
      parameters:
        - $ref: '#/components/parameters/ServerPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [uri]
              properties:
                uri:
                  type: string
                  description: Resource URI
      responses:
        '200':
          description: Resource content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceContent'
  /v1/mcp/{server}/prompts/list:
    post:
      tags: [MCP]
      summary: List prompts
      description: List all prompts available on the specified MCP server.
      operationId: listPrompts
      parameters:
        - $ref: '#/components/parameters/ServerPath'
      responses:
        '200':
          description: List of prompts
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Prompt'
  /v1/mcp/{server}/prompts/get:
    post:
      tags: [MCP]
      summary: Get a prompt
      description: Get a prompt with rendered messages from the specified MCP server.
      operationId: getPrompt
      parameters:
        - $ref: '#/components/parameters/ServerPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: Prompt name
                arguments:
                  type: object
                  description: Prompt arguments
      responses:
        '200':
          description: Prompt messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/PromptMessage'

  # Traces
...
  /v1/approvals:
    get:
      tags: [Safety]
      summary: List tool approvals
      description: List tool approval requests.
      operationId: listApprovals
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, denied]
      responses:
        '200':
          description: List of approvals
          content:
            application/json:
              schema:
                type: object
                properties:
                  approvals:
                    type: array
                    items:
                      $ref: '#/components/schemas/ToolApproval'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  parameters:
    ServerPath:
      name: server
      in: path
      required: true
      schema:
        type: string
      description: MCP server name
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    ToolDefinition:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        inputSchema:
          type: object
    ToolCallResult:
      type: object
      properties:
        content:
          description: Tool output
        isError:
          type: boolean
        metadata:
          type: object
        traceId:
          type: string
        spanId:
          type: string
        durationMs:
          type: integer
        cost:
          type: number
    Resource:
      type: object
      properties:
        uri:
          type: string
        name:
          type: string
        description:
          type: string
        mimeType:
          type: string
    ResourceContent:
      type: object
      properties:
        uri:
          type: string
        mimeType:
          type: string
        text:
          type: string
        blob:
          type: string
          description: Base64 encoded binary content
    Prompt:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        arguments:
          type: array
          items:
            type: object
    PromptMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          description: Message content
...
    ToolApproval:
      type: object
      properties:
        id:
          type: string
        mcpServer:
          type: string
        toolName:
          type: string
        teamId:
          type: string
        status:
          type: string
          enum: [pending, approved, denied]
        requestedBy:
          type: string
        reviewedBy:
          type: string
        reviewedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...